generator: add namespace template variables
Add template variables for namespace associated to project.

The namespace is an abbreviation of the name of the project, using the
first letter of no more than the first three words in the name, i.e.
"foo-bar-quux" or "Foo Bar Quux" becomes "Fbq", but "foobar" stays
"foobar".

It is possible to override the computed value using the "nsname" key.